promote_action

+++ TrackQueueActor.py
@@ -21,7 +21,7 @@ class Player(pykka.ThreadingActor):
         cmd = "mpg123 %s" % track
         p = subprocess.Popen(cmd, shell=True)
 
-        self.queue_actor.tell({'command':'playing_process', "p":p})
+        self.queue_actor.tell({'command': 'playing_process', "p": p})
 
         p.wait()
 
@@ -43,11 +43,10 @@ class Player(pykka.ThreadingActor):
     def check_queue(self):
         to_play = self.queue_actor.ask({'command': 'pop_play'})
         if to_play is not None:
-            if to_play.get('skip'):
+            if to_play.get('action') == 'skip' or to_play.get('action') == 'promote':
                 self.check_queue()
             else:
                 self.play_and_report(to_play)
-
 
     def on_receive(self, message):
         if message.get('command') == 'check':
@@ -80,12 +79,13 @@ class Downloader(pykka.ThreadingActor):
             self.queue_actor.tell({'command': 'downloaded', 'track': track, 'file': mp3_track})
         except Exception as ex:
             print ex
-        finally: self.check_queue()
+        finally:
+            self.check_queue()
 
     def check_queue(self):
         to_download = self.queue_actor.ask({'command': 'pop_download'})
         if to_download is not None:
-            if to_download.get('skip'):
+            if to_download.get('action') == 'skip' or to_download.get('action') == 'promote':
                 self.check_queue()
             else:
                 self.download(to_download)
@@ -102,6 +102,10 @@ class TrackQueueActor(pykka.ThreadingActor):
 
         self.download_queue = Queue()
         self.player_queue = Queue()
+
+        self.download_queue_promoted = Queue()
+        self.player_queue_promoted = Queue()
+
         self.count = 0
         self.player = Player.start(mqtt_actor, self.actor_ref)
         self.downloader = Downloader.start(mqtt_actor, self.actor_ref)
@@ -110,6 +114,7 @@ class TrackQueueActor(pykka.ThreadingActor):
         self.playing = None
 
         self.skip_current_download = False
+        self.promote_current_download = False
         self.p = None
 
     # def check_download(self):
@@ -120,27 +125,34 @@ class TrackQueueActor(pykka.ThreadingActor):
     #         track['count'] = self.count
     #         self.downloader.tell({'command':'download', 'track': track})
 
-    def on_skip(self, orig):
+    def on_move_action(self, orig, action):
         track = None
         if self.downloading is not None and self.downloading.get('orig') == orig:
-            self.skip_current_download = True
+            if action == 'skip':
+                self.skip_current_download = True
+                self.promote_current_download = False
+            else:
+                self.skip_current_download = False
+                self.promote_current_download = True
             track = self.downloading
         elif self.playing is not None and self.playing.get('track').get('orig') == orig:
             track = self.playing.get('track')
-            if self.p is not None:
+            if action == 'skip' and self.p is not None:
                 self.p.terminate()
         else:
             for qp in self.player_queue.queue:
                 if qp.get('track').get('orig') == orig:
-                    qp['skip'] = True
+                    self.player_queue_promoted.put(qp.copy())
+                    qp['action'] = action
                     track = qp.get('track')
             for qd in self.download_queue.queue:
                 if qd.get('orig') == orig:
-                    qd['skip'] = True
+                    self.download_queue_promoted.put(qd.copy())
+                    qd['action'] = action
                     track = qd.get('track')
         if track is not None:
-            track['message'] = 'skip'
-            self.mqtt_actor.tell({'command': 'update_track_status', 'status': 'skip', 'track': track})
+            track['message'] = action
+            self.mqtt_actor.tell({'command': 'update_track_status', 'status': action, 'track': track})
 
     def on_receive(self, message):
         if message.get('command') == 'track':
@@ -148,31 +160,45 @@ class TrackQueueActor(pykka.ThreadingActor):
             self.download_queue.put(message.get('track'))
             self.downloader.tell({'command': 'check'})
         elif message.get('command') == 'pop_download':
-            self.downloading = None if self.download_queue.empty() else self.download_queue.get(block=False)
+            self.downloading = None if self.download_queue_promoted.empty() else self.download_queue_promoted.get(
+                block=False)
+            if self.downloading is None:
+                self.downloading = None if self.download_queue.empty() else self.download_queue.get(block=False)
             if self.downloading is not None:
                 self.count += 1
                 self.downloading['count'] = self.count
             return self.downloading
         elif message.get('command') == 'pop_play':
-            self.playing = None if self.player_queue.empty() else self.player_queue.get(block=False)
+            self.playing = None if self.player_queue_promoted.empty() else self.player_queue_promoted.get(block=False)
+            if self.playing is None:
+                self.playing = None if self.player_queue.empty() else self.player_queue.get(block=False)
             return self.playing
         # elif message.get('command') == 'check_download':
         #     self.check_download()
         elif message.get('command') == 'startup':
             self.player.tell(message)
         elif message.get('command') == 'skip':
-            self.on_skip(message.get('orig'))
+            self.on_move_action(message.get('orig'), 'skip')
+        elif message.get('command') == 'promote':
+            self.on_move_action(message.get('orig'), 'promote')
         # elif message.get('command') == 'check_player':
         #     self.check_player()
         elif message.get('command') == 'downloaded':
             if self.skip_current_download:
                 self.skip_current_download = False
             else:
-                print ('add track ' + message.get('file') + ' to play queue')
+                if self.promote_current_download:
+                    self.promote_current_download = False
+                    self.player_queue_promoted.put(message)
+                else:
+                    self.player_queue.put(message)
+
                 track = message.get('track')
                 track['message'] = 'queue'
                 self.mqtt_actor.tell({'command': 'update_track_status', 'status': 'queue', 'track': track})
-                self.player_queue.put(message)
                 self.player.tell({'command': 'check'})
+
+                print ('add track ' + message.get('file') + ' to play queue')
+
         elif message.get('command') == 'playing_process':
             self.p = message.get('p')
